feat(static-web): implement static OG image generation service and add font support
- Updated Dockerfile to use Alpine variant for Node.js and added Perl installation.
- Introduced StaticOgService for generating Open Graph images with customizable templates.
- Added Geist-Medium font for rendering OG images.
- Updated TypeScript configuration to include JSX support and adjusted paths.
- Created type declarations for assets and Vercel OG module.
- Added new static assets for OG image generation.
