+++ be/apps/core/src/types/assets.d.ts
@@ -1,0 +1,4 @@
+declare module '*.ttf?url' {
+  const url: string
+  export default url
+}
